Partial fix of a jpx inclusion tag tree bug
The current code only works when every code-block includes at least some
coding pass in layer 0. Otherwise, the tag tree will have x preceding 0,
where x is the number of layers before the code-block is first included
(See section B.10.2 and B.10.4 of ITU-T Rec. T.800) and the order of tag
tree will no be raster if the codeblock are not all introduced at the
same layer.

Currently, file that have packets of layer 0 that don't contain data for
every codeblocks appears severely corrupted as everything read beyond that
point is misaligned.

This commit purge the preceding zeros when the inclusionTree is constructed.
This will only work when all codeblocks are first included in the same
packet.

An error is thrown if the number of preceding zeros does not match the current
layer. It is unrecoverable in most cases.